ai_service/requirements-stable.txt
# For Python 3.8-3.11 (stable/recommended)
tensorflow-cpu==2.15.0
pillow==10.0.0
numpy==1.24.4
flask==2.3.3
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
